+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Fix reactivity in Leader component
 - Fix events triggered by Accordion component
 - Fix text wrapping for Firefox in Position component
+- Fix Scrollspy triggering before other components have been applied at least once
 
 ## 3.0.0 beta 34 (November 7, 2017)